// types/movie.ts
export interface Movie {
  id: number;
  title: string;
  overview: string;
  poster_path: string;
  backdrop_path: string;
  release_date: string;
  vote_average: number;
  genre_ids: number[];
  media_type?: string;
  video_id?: string;
}

export interface Genre {
  id: number;
  name: string;
}

export interface MovieRow {
  title: string;
  movies: Movie[];
}